+++ src/init.lua
@@ -7,6 +7,8 @@ export type Color = Color.Color
 export type Gradient = Gradient.Gradient
 export type GradientKeypoint = Gradient.GradientKeypoint
 
+---
+
 return {
     new = Color.new,
     random = Color.random,
@@ -35,4 +37,42 @@ return {
     gradient = Gradient.new,
     gradientFromColors = Gradient.fromColors,
     gradientFromColorSequence = Gradient.fromColorSequence,
+
+    fromBrickColor = Color.fromBrickColor,
+    fromCMYK = Color.fromCMYK,
+    fromColor3 = Color.fromColor3,
+    fromHex = Color.fromHex,
+    fromHSB = Color.fromHSB,
+    fromHSL = Color.fromHSL,
+    fromHWB = Color.fromHWB,
+    fromLab = Color.fromLab,
+    fromLChab = Color.fromLChab,
+    fromLChuv = Color.fromLChuv,
+    fromLuv = Color.fromLuv,
+    fromNumber = Color.fromNumber,
+    fromRGB = Color.fromRGB,
+    fromTemperature = Color.fromTemperature,
+    fromXYZ = Color.fromXYZ,
+
+    toBrickColor = Color.toBrickColor,
+    toCMYK = Color.toCMYK,
+    toColor3 = Color.toColor3,
+    toHex = Color.toHex,
+    toHSB = Color.toHSB,
+    toHSL = Color.toHSL,
+    toHWB = Color.toHWB,
+    toLab = Color.toLab,
+    toLChab = Color.toLChab,
+    toLChuv = Color.toLChuv,
+    toLuv = Color.toLuv,
+    toNumber = Color.toNumber,
+    toRGB = Color.toRGB,
+    toTemperature = Color.toTemperature,
+    toXYZ = Color.toXYZ,
+
+    fromHSV = Color.fromHSV,
+    fromLCh = Color.fromLCh,
+
+    toHSV = Color.toHSV,
+    toLCh = Color.toLCh,
 }